Introduce aggregation helper for partitioned tables
