+++ custom/icds_reports/utils/aggregation_helpers/distributed/base.py
@@ -1,11 +1,14 @@
 import hashlib
 import logging
 
+from django.db import transaction
+
 from corehq.apps.userreports.models import (
     StaticDataSourceConfiguration,
     get_datasource_config,
 )
 from corehq.apps.userreports.util import get_table_name
+from corehq.sql_db.routers import db_for_read_write
 from custom.icds_reports.const import DASHBOARD_DOMAIN
 from custom.icds_reports.utils.aggregation_helpers import (
     AggregationHelper,
@@ -129,3 +132,72 @@ class StateBasedAggregationPartitionedHelper(BaseICDSAggregationDistributedHelpe
             "month_string": month_string,
             "state_id": self.state_id
         }
+
+
+class AggregationPartitionedHelper(BaseICDSAggregationDistributedHelper):
+    """Helper for tables that reside on Citus master and are partitioned into one tables per month
+    """
+    staging_tablename = None
+    base_tablename = None
+    monthly_tablename = None
+    model = None
+
+    def __init__(self, month):
+        self.month = transform_day_to_month(month)
+
+    def aggregate(self, cursor):
+        staging_queries = self.staging_queries()
+        update_queries = self.update_queries()
+        rollup_queries = [self.rollup_query(i) for i in range(4, 0, -1)]
+
+        logger.info(f"Creating staging table for {self.helper_key}")
+        cursor.execute(f"DROP TABLE IF EXISTS {self.staging_tablename}")
+        cursor.execute(f"CREATE UNLOGGED TABLE {self.staging_tablename} (LIKE {self.base_tablename})")
+
+        logger.info(f"Inserting inital data into staging table for {self.helper_key}")
+        for staging_query, params in staging_queries:
+            cursor.execute(staging_query, params)
+
+        logger.info(f"Updating data into staging table for {self.helper_key}")
+        for query, params in update_queries:
+            cursor.execute(query, params)
+
+        logger.info(f"Rolling up data into staging table for {self.helper_key}")
+        for query in rollup_queries:
+            cursor.execute(query)
+
+        logger.info(f"Creating new table for {self.helper_key} {self.month}")
+        cursor.execute(f"""
+        CREATE TABLE IF NOT EXISTS "{self.monthly_tablename}" (
+            CHECK (month = DATE '{self.month.strftime('%Y-%m-01')}')
+        )
+        INHERITS ({self.base_tablename})
+        """)
+
+        logger.info(f"Creating indexes for {self.helper_key} {self.month}")
+        for index_query in self.indexes():
+            cursor.execute(index_query)
+
+        db_alias = db_for_read_write(self.model)
+        with transaction.atomic(using=db_alias):
+            logger.info(f"Dropping legacy tables for {self.helper_key} {self.month}")
+            for i in range(1, 6):
+                cursor.execute(f'DROP TABLE IF EXISTS "{self._legacy_tablename_func(i)}"')
+
+            logger.info(f"Removing current data for {self.helper_key} {self.month}")
+            cursor.execute(f'DELETE FROM "{self.monthly_tablename}"')
+
+            logger.info(f"Inserting new data for {self.helper_key} {self.month}")
+            cursor.execute(f"""
+            INSERT INTO "{self.monthly_tablename}" (
+                SELECT * FROM "{self.staging_tablename}"
+                ORDER BY aggregation_level, state_id, district_id, block_id, supervisor_id, awc_id
+            )
+            """)
+
+        logger.info(f"Dropping staging table for {self.helper_key} {self.month}")
+        cursor.execute(f"DROP TABLE IF EXISTS {self.staging_tablename}")
+        logger.info(f"Finished aggregation for {self.helper_key} {self.month}")
+
+    def _legacy_tablename_func(self, agg_level):
+        return "{}_{}_{}".format(self.base_tablename, self.month.strftime("%Y-%m-%d"), agg_level)
